requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1